move code pieces to appropriate files
