+++ plugin/evm/atomic/state/atomic_trie_iterator.go
@@ -8,12 +8,42 @@ import (
 	"fmt"
 
 	"github.com/ava-labs/avalanchego/chains/atomic"
+	avalancheatomic "github.com/ava-labs/avalanchego/chains/atomic"
 	"github.com/ava-labs/avalanchego/codec"
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/utils/wrappers"
 
 	"github.com/ava-labs/coreth/trie"
 )
+
+var _ AtomicTrieIterator = &atomicTrieIterator{}
+
+// AtomicTrieIterator is a stateful iterator that iterates the leafs of an AtomicTrie
+type AtomicTrieIterator interface {
+	// Next advances the iterator to the next node in the atomic trie and
+	// returns true if there are more leaves to iterate
+	Next() bool
+
+	// Key returns the current database key that the iterator is iterating
+	// returned []byte can be freely modified
+	Key() []byte
+
+	// Value returns the current database value that the iterator is iterating
+	Value() []byte
+
+	// BlockNumber returns the current block number
+	BlockNumber() uint64
+
+	// BlockchainID returns the current blockchain ID at the current block number
+	BlockchainID() ids.ID
+
+	// AtomicOps returns a map of blockchainIDs to the set of atomic requests
+	// for that blockchainID at the current block number
+	AtomicOps() *avalancheatomic.Requests
+
+	// Error returns error, if any encountered during this iteration
+	Error() error
+}
 
 // atomicTrieIterator is an implementation of types.AtomicTrieIterator that serves
 // parsed data with each iteration
@@ -27,7 +57,7 @@ type atomicTrieIterator struct {
 	err          error            // error if any has occurred
 }
 
-func NewAtomicTrieIterator(trieIterator *trie.Iterator, codec codec.Manager) AtomicTrieIterator {
+func NewAtomicTrieIterator(trieIterator *trie.Iterator, codec codec.Manager) *atomicTrieIterator {
 	return &atomicTrieIterator{trieIterator: trieIterator, codec: codec}
 }
